requirements.txt
flet==0.22.*
pytz 
python-dateutil
aiohttp
matplotlib